PromQL: rename PromqlAttributesTranslationContext
A pure rename to TranslationContext, with its test. The class is the
translation's shared context - the label header the translator pushes
down and lifts up and the name mapping between PromQL labels and ES|QL
attributes - not a bag of attribute helpers, and the follow-up commits
build the header model on it under that name. No behavior change.
